fix: service info error on build
